feat: refactor login page screen
