+++ view/packages/components/otp-login-form.tsx
@@ -1,13 +1,14 @@
 import { cn } from '@/lib/utils';
 import { Button } from '@nixopus/ui';
-import { Card, CardContent, CardHeader, CardTitle, CardDescription } from '@nixopus/ui';
+import { Card, CardContent } from '@nixopus/ui';
 import { Input } from '@nixopus/ui';
 import { Label } from '@nixopus/ui';
 import { Alert, AlertDescription } from '@nixopus/ui';
 import { OTPInput } from '@nixopus/ui';
-import nixopusLogo from '@/public/nixopus_logo_transparent.png';
+import nixopusLogo from '@/public/logo_white.png';
 import { useTranslation } from '@/packages/hooks/shared/use-translation';
 import { useOtpLoginForm } from '@/packages/hooks/auth/use-otp-login-form';
+import { Mail } from 'lucide-react';
 
 export interface OtpLoginFormProps {
   email: string;
@@ -37,27 +38,28 @@ export function OtpLoginForm({ ...props }: OtpLoginFormProps) {
         <CardContent className="grid p-0 md:grid-cols-2">
           <div className="p-6 md:p-8">
             <div className="flex flex-col gap-6">
-              <CardHeader className="p-0 text-center">
-                <CardTitle className="text-2xl">
-                  {props.otpSent ? t('auth.otpLogin.verifyTitle') : t('auth.otpLogin.title')}
-                </CardTitle>
-                <CardDescription>
-                  {props.otpSent
-                    ? t('auth.otpLogin.verifyDescription')
-                    : t('auth.otpLogin.description')}
-                </CardDescription>
-              </CardHeader>
+              {props.otpSent && (
+                <div className="flex flex-col items-center text-center">
+                  <h1 className="text-2xl font-bold">{t('auth.otpLogin.verifyTitle')}</h1>
+                  <p className="text-muted-foreground text-balance">
+                    {t('auth.otpLogin.verifyDescription')}
+                  </p>
+                </div>
+              )}
               {!props.otpSent && (
                 <div className="grid gap-3">
-                  <Label htmlFor="email">{t('auth.email')}</Label>
-                  <Input
-                    id="email"
-                    type="email"
-                    placeholder={t('auth.login.emailPlaceholder')}
-                    required
-                    value={props.email}
-                    onChange={props.handleEmailChange}
-                  />
+                  <div className="relative">
+                    <Mail className="absolute left-3 top-1/2 h-4 w-4 -translate-y-1/2 text-muted-foreground" />
+                    <Input
+                      id="email"
+                      type="email"
+                      className="pl-10"
+                      placeholder={t('auth.login.emailPlaceholder')}
+                      required
+                      value={props.email}
+                      onChange={props.handleEmailChange}
+                    />
+                  </div>
                   {emailError && (
                     <Alert variant="destructive">
                       <AlertDescription className="text-xs !text-red-600 font-medium">
@@ -120,7 +122,7 @@ export function OtpLoginForm({ ...props }: OtpLoginFormProps) {
             <img
               src={nixopusLogo.src}
               alt="Nixopus Logo"
-              className="absolute inset-0 h-full w-full object-cover"
+              className="absolute inset-0 h-full w-full object-contain p-8"
             />
           </div>
         </CardContent>
